+++ .github/workflows/ci-testing.yml
@@ -60,7 +60,7 @@ jobs:
 
     - name: Tests
       run: |
-        coverage run --source src,detoxify -m py.test src tests -v --junitxml=junit/test-results-${{ runner.os }}-${{ matrix.python-version }}.xml
+        coverage run --source detoxify -m py.test detoxify tests -v --junitxml=junit/test-results-${{ runner.os }}-${{ matrix.python-version }}.xml
 
     - name: Statistics
       if: success()
