--- model_eval/utils.py
@@ -1,14 +1,0 @@
-import numpy as np
-from sklearn.metrics import roc_auc_score
-
-
-def compute_auc(y_true, y_pred):
-    try:
-        return roc_auc_score(y_true, y_pred)
-    except ValueError:
-        return np.nan
-
-
-def compute_subgroup_auc(df, subgroup, label, model_name):
-    subgroup_examples = df[df[subgroup]]
-    return compute_auc(subgroup_examples[label], subgroup_examples[model_name])
