+++ scripts/train.py
@@ -1,14 +1,14 @@
 import argparse
 import json
-import os
 
 import pytorch_lightning as pl
-import src.data_loaders as module_data
 import torch
 from pytorch_lightning.callbacks import ModelCheckpoint
-from src.utils import get_model_and_tokenizer
 from torch.nn import functional as F
 from torch.utils.data import DataLoader
+
+import detoxify.utils.data_loaders as module_data
+from detoxify.utils.extra import get_model_and_tokenizer
 
 
 class ToxicClassifier(pl.LightningModule):
